+++ packages/lint/src/index.ts
@@ -1,11 +1,13 @@
 import buttonRules from './rules/button-rules.js';
 import containerRules from './rules/container-rules.js';
+import imageRules from './rules/image-rules.js';
 import linkRules from './rules/link-rules.js';
 import textRules from './rules/text-rules.js';
 
 const rules = {
   ...buttonRules.rules,
   ...containerRules.rules,
+  ...imageRules.rules,
   ...linkRules.rules,
   ...textRules.rules,
 };
@@ -15,6 +17,7 @@ const configs = {
     rules: {
       ...buttonRules.configs.recommended.rules,
       ...containerRules.configs.recommended.rules,
+      ...imageRules.configs.recommended.rules,
       ...linkRules.configs.recommended.rules,
       ...textRules.configs.recommended.rules,
     },
@@ -33,6 +36,8 @@ export type {
   ButtonTagAnalysis,
   ContainerClassInfo,
   ContainerTagAnalysis,
+  ImageClassInfo,
+  ImageTagAnalysis,
   LinkClassInfo,
   LinkTagAnalysis,
   TextComponentInfo,