+++ sandbox/vue-library/src/index.ts
@@ -1,5 +1,6 @@
 import CedarButtonFixture from './components/CedarButtonFixture.vue';
 import CedarContainerFixture from './components/CedarContainerFixture.vue';
+import CedarImageFixture from './components/CedarImageFixture.vue';
 import CedarLinkFixture from './components/CedarLinkFixture.vue';
 import CedarTextFixture from './components/CedarTextFixture.vue';
 import IconUser from './components/IconUser.vue';
@@ -7,6 +8,7 @@ import IconUser from './components/IconUser.vue';
 export {
   CedarButtonFixture,
   CedarContainerFixture,
+  CedarImageFixture,
   CedarLinkFixture,
   CedarTextFixture,
   IconUser,