+++ stories/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @rei/cedar-stories Changelog
 
+## 0.1.1-alpha.0 - 2026-01-29
+
+- Added image stories and updated fixtures to use Cedar brand imagery.
+
 ## 0.1.0-alpha.0 - 2026-01-27
 
 - Added Storybook stories for button, container, and link variants.